TalentDataCrawl/requirements.txt
Scrapy==1.8.0
beautifulsoup4==4.8.1
elasticsearch==7.1.0
gensim==3.8.1
lxml=4.4.1
pyahocorasick==1.4.0
pyvi==0.0.9.7
requests==2.22.0
scikit-learn==0.20.3
tensorflow==2.0.0
urllib3==1.25.7
underthesea==1.1.17
